Upgrade to Android Gradle Plugin 3.5.3 and Gradle 5.4.1.
Gradle's API for dependency transforms changed.
Apply RoboJavaModulePlugin in a more idiomatic way.
