--- buildSrc/src/main/groovy/RoboJavaModulePlugin.groovy
@@ -1,191 +1,0 @@
-import org.gradle.api.JavaVersion
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.artifacts.maven.MavenDeployment
-import org.gradle.api.tasks.bundling.Jar
-import org.gradle.api.tasks.compile.JavaCompile
-
-class RoboJavaModulePlugin implements Plugin<Project> {
-    Boolean deploy = false;
-
-    Closure doApply = {
-        apply plugin: "java-library"
-        apply plugin: "net.ltgt.errorprone"
-
-        apply plugin: org.robolectric.gradle.AarDepsPlugin
-
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-
-        project.dependencies {
-            errorprone("com.google.errorprone:error_prone_core:$errorproneVersion")
-            errorproneJavac("com.google.errorprone:javac:$errorproneJavacVersion")
-        }
-
-        tasks.withType(JavaCompile) { task ->
-            sourceCompatibility = JavaVersion.VERSION_1_8
-            targetCompatibility = JavaVersion.VERSION_1_8
-
-            // Show all warnings except boot classpath
-            configure(options) {
-                if (System.properties["lint"] != null && System.properties["lint"] != "false") {
-                    compilerArgs << "-Xlint:all"        // Turn on all warnings
-                }
-                compilerArgs << "-Xlint:-options"       // Turn off "missing" bootclasspath warning
-                encoding = "utf-8"                      // Make sure source encoding is UTF-8
-            }
-
-            def noRebuild = System.getenv('NO_REBUILD') == "true"
-            if (noRebuild) {
-                println "[NO_REBUILD] $task will be skipped!"
-                task.outputs.upToDateWhen { true }
-                task.onlyIf { false }
-            }
-        }
-
-        ext.mavenArtifactName = project.path.substring(1).split(/:/).join("-")
-
-        task('provideBuildClasspath', type: ProvideBuildClasspathTask) {
-            File outDir = project.sourceSets['test'].output.resourcesDir
-            outFile = new File(outDir, 'robolectric-deps.properties')
-        }
-
-        tasks['test'].dependsOn provideBuildClasspath
-
-        test {
-            exclude "**/*\$*" // otherwise gradle runs static inner classes like TestRunnerSequenceTest$SimpleTest
-
-            // TODO: DRY up code with AndroidProjectConfigPlugin...
-            testLogging {
-                exceptionFormat "full"
-                showCauses true
-                showExceptions true
-                showStackTraces true
-                showStandardStreams true
-                events = ["failed", "skipped"]
-            }
-
-            minHeapSize = "1024m"
-            maxHeapSize = "4096m"
-
-            if (System.env['GRADLE_MAX_PARALLEL_FORKS'] != null) {
-                maxParallelForks = Integer.parseInt(System.env['GRADLE_MAX_PARALLEL_FORKS'])
-            }
-
-            def forwardedSystemProperties = System.properties
-                    .findAll { k,v -> k.startsWith("robolectric.") }
-                    .collect { k,v -> "-D$k=$v" }
-            jvmArgs = ["-XX:MaxPermSize=1024m"] + forwardedSystemProperties
-
-            doFirst {
-                if (!forwardedSystemProperties.isEmpty()) {
-                    println "Running tests with ${forwardedSystemProperties}"
-                }
-            }
-        }
-
-        if (owner.deploy) {
-            project.apply plugin: "signing"
-            project.apply plugin: "maven"
-            project.apply plugin: 'ch.raffael.pegdown-doclet'
-
-            task('sourcesJar', type: Jar, dependsOn: classes) {
-                classifier "sources"
-                from sourceSets.main.allJava
-            }
-
-            javadoc {
-                failOnError = false
-                source = sourceSets.main.allJava
-                options.noTimestamp = true
-                options.header = "<ul class=\"navList\"><li>Robolectric $thisVersion | <a href=\"/\">Home</a></li></ul>"
-                options.footer = "<ul class=\"navList\"><li>Robolectric $thisVersion | <a href=\"/\">Home</a></li></ul>"
-                options.bottom = "<link rel=\"stylesheet\" href=\"https://robolectric.org/assets/css/main.css\">"
-                options.version = thisVersion
-            }
-
-            task('javadocJar', type: Jar, dependsOn: javadoc) {
-                classifier "javadoc"
-                from javadoc.destinationDir
-            }
-
-            signing {
-                required { !version.endsWith("SNAPSHOT") && gradle.taskGraph.hasTask("uploadArchives") }
-                sign configurations.archives
-            }
-
-            def skipJavadoc = System.getenv('SKIP_JAVADOC') == "true"
-            artifacts {
-                archives jar
-                archives sourcesJar
-                if (!skipJavadoc) {
-                    archives javadocJar
-                }
-            }
-
-            // for maven local install:
-            archivesBaseName = mavenArtifactName
-
-            uploadArchives {
-                repositories {
-                    mavenDeployer {
-                        pom.artifactId = mavenArtifactName
-                        pom.project {
-                            name project.name
-                            description = "An alternative Android testing framework."
-                            url = "http://robolectric.org"
-
-                            licenses {
-                                license {
-                                    name "The MIT License"
-                                    url "https://opensource.org/licenses/MIT"
-                                }
-                            }
-
-                            scm {
-                                url "git@github.com:robolectric/robolectric.git"
-                                connection "scm:git:git://github.com/robolectric/robolectric.git"
-                                developerConnection "scm:git:https://github.com/robolectric/robolectric.git"
-                            }
-
-                            developers {
-                                developer {
-                                    name "Christian Williams"
-                                    email "christianw@google.com"
-                                    organization = "Google Inc."
-                                    organizationUrl "http://google.com"
-                                }
-
-                                developer {
-                                    name "Jonathan Gerrish"
-                                    email "jongerrish@google.com"
-                                    organization = "Google Inc."
-                                    organizationUrl "http://google.com"
-                                }
-                            }
-                        }
-
-                        def url = project.version.endsWith("-SNAPSHOT") ?
-                                "https://oss.sonatype.org/content/repositories/snapshots" :
-                                "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
-                        repository(url: url) {
-                            authentication(
-                                    userName: System.properties["sonatype-login"] ?: System.env['sonatypeLogin'],
-                                    password: System.properties["sonatype-password"] ?: System.env['sonatypePassword']
-                            )
-                        }
-
-                        beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
-                    }
-                }
-            }
-        }
-    }
-
-    @Override
-    void apply(Project project) {
-        doApply.delegate = project
-        doApply.resolveStrategy = Closure.DELEGATE_ONLY
-        doApply()
-    }
-}
