+++ integration_tests/mockito/build.gradle
@@ -1,4 +1,4 @@
-apply plugin: RoboJavaModulePlugin
+apply plugin: org.robolectric.gradle.RoboJavaModulePlugin
 
 dependencies {
     api project(":robolectric")
