+++ pluginapi/build.gradle
@@ -1,6 +1,5 @@
-new RoboJavaModulePlugin(
-        deploy: true
-).apply(project)
+apply plugin: org.robolectric.gradle.RoboJavaModulePlugin
+apply plugin: org.robolectric.gradle.DeployedRoboJavaModulePlugin
 
 dependencies {
     compileOnly 'com.google.code.findbugs:jsr305:3.0.2'
