+++ shadows/framework/build.gradle
@@ -1,6 +1,5 @@
-new RoboJavaModulePlugin(
-        deploy: true
-).apply(project)
+apply plugin: org.robolectric.gradle.RoboJavaModulePlugin
+apply plugin: org.robolectric.gradle.DeployedRoboJavaModulePlugin
 
 apply plugin: ShadowsPlugin
 
@@ -43,7 +42,7 @@ dependencies {
     api project(":shadowapi")
     api project(":utils")
     api project(":utils:reflector")
-    api "androidx.test:monitor:1.3.0-alpha04"
+    api "androidx.test:monitor:1.3.0-alpha04@aar"
 
     compileOnly "com.google.code.findbugs:jsr305:3.0.2"
     api "com.almworks.sqlite4java:sqlite4java:0.282"
